requirements.txt
ultralytics
opencv-python
pillow
kagglehub
seaborn
pandas
matplotlib
